+++ cli/src/commands/subgraph/commands/publish.ts
@@ -8,7 +8,7 @@ import pc from 'picocolors';
 import { EnumStatusCode } from '@wundergraph/cosmo-connect/dist/common/common_pb';
 import { parseGraphQLSubscriptionProtocol, splitLabel } from '@wundergraph/cosmo-shared';
 import { BaseCommandOptions } from '../../../core/types/types.js';
-import { baseHeaders } from '../../../core/config.js';
+import { getBaseHeaders } from '../../../core/config.js';
 
 export default (opts: BaseCommandOptions) => {
   const command = new Command('publish');
@@ -89,7 +89,7 @@ export default (opts: BaseCommandOptions) => {
         unsetLabels: !!options.unsetLabels,
       },
       {
-        headers: baseHeaders,
+        headers: getBaseHeaders(),
       },
     );