+++ router/CHANGELOG.md
@@ -4,6 +4,24 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.79.2](https://github.com/wundergraph/cosmo/compare/router@0.79.1...router@0.79.2) (2024-04-06)
+
+### Bug Fixes
+
+* remove unused host-cpu metric ([#703](https://github.com/wundergraph/cosmo/issues/703)) ([a664631](https://github.com/wundergraph/cosmo/commit/a664631477d8b30b2448f8284db257f4d33446e6)) (@StarpTech)
+
+## [0.79.1](https://github.com/wundergraph/cosmo/compare/router@0.79.0...router@0.79.1) (2024-04-04)
+
+### Bug Fixes
+
+* create individual metrics per error code ([#693](https://github.com/wundergraph/cosmo/issues/693)) ([d63d087](https://github.com/wundergraph/cosmo/commit/d63d087145d7ebeeba0673241e8f902c4008679b)) (@StarpTech)
+
+# [0.79.0](https://github.com/wundergraph/cosmo/compare/router@0.78.0...router@0.79.0) (2024-04-03)
+
+### Features
+
+* subgraph error handling and instrumentation ([#675](https://github.com/wundergraph/cosmo/issues/675)) ([55a1215](https://github.com/wundergraph/cosmo/commit/55a1215f5c1ee417a7287c89520b1540c5066b82)) (@StarpTech)
+
 # [0.78.0](https://github.com/wundergraph/cosmo/compare/router@0.77.0...router@0.78.0) (2024-03-25)
 
 ### Features